added stories for modal
