+++ src/.storybook/config.js
@@ -2,7 +2,7 @@ import { configure } from '@kadira/storybook';
 
 function loadStories() {
   require('../styles/index.css');
-  require('../components/stories/button');
+  require('../components/stories/');
   // require as many stories as you need.
 }
