+++ src/components/stories/index.js
@@ -1,0 +1,2 @@
+import './button';
+import './modal';
