+++ src/components/stories/info.js
@@ -1,0 +1,20 @@
+import React from 'react';
+
+const Info = ({ children }) => {
+  return (
+    <p className="p2 mt2 border rounded"
+      style={{
+        color: '#8a6d3b',
+        backgroundColor: '#fcf8e3',
+        borderColor: '#faebcc',
+      }}>
+      { children }
+    </p>
+  );
+};
+
+Info.propTypes = {
+  children: React.PropTypes.node.isRequired,
+};
+
+export default Info;
